% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CUSUM-filter.R
\name{CUSUM_Price}
\alias{CUSUM_Price}
\title{CUSUM Filter for Price}
\usage{
CUSUM_Price(data, h)
}
\arguments{
\item{data}{data in form of bar (tick bar, time bar, volume bar ...)}

\item{h}{threshold value}
}
\value{
an index of data exceeding threshold value h
}
\description{
CUSUM Filter for Price
}
\examples{
CUSUM_Price(data1, 2)
}
