% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mladv-package.R
\docType{package}
\name{mladv-package}
\alias{mladv-package}
\title{Package mladv}
\description{
The goal of the package is to provide tools to clean, convert and analyze
financial data
}
\section{Help}{

For a quick help see the README file.
For more details see the vignette
}

\keyword{internal}
