% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timebar.R
\name{timebar}
\alias{timebar}
\title{Time bar transformations from raw data}
\usage{
timebar(data, type)
}
\arguments{
\item{data}{a data.frame with follwing headers name: "Time", "Price", "Size", "cum_volume" and "ppt"}

\item{type}{type of time bar from "minute", "hour" or "day"}
}
\value{
a data.frame in the form OHLC with ending date for each bar.
}
\description{
Function to transform raw transactions data into a time bar for a specific timeframe.
}
\examples{
transform_data <- bar_time(SSI_data, "minute")
}
