% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volumebarr.r
\name{volumebarr}
\alias{volumebarr}
\title{Volume bar transformations from raw data}
\usage{
volumebarr(data, vol)
}
\arguments{
\item{data}{a data.frame with follwing headers name: "Time", "Price", "Size", "cum_volume" and "ppt"}

\item{vol}{integer number for volume level}
}
\value{
a data.frame in the form OHLC with ending date for each bar.
}
\description{
Volume bar transformations from raw data
}
\examples{
tranform_data <- volumebarr(SSI_data, 5000)
}
